Selection soorting

+++ 2-selection_sort.c
@@ -8,7 +8,7 @@ void switch_(int *, int, int);
  */
 void selection_sort(int *array, size_t size)
 {
-	unsigned int i = 0, j = 0, min;
+	size_t i = 0, j = 0, min;
 
 	if (array == NULL  || size == 0)
 		return;
